Move EventsFileSerialization code and test code to own classes
